Document "ignored" blocks
If using a block type other than `js` or `css` the block will be "ignored" and replaced by an empty string.  This behaviour is enabled by fixes in #303 and I believe it should be documented because it is incredibly useful for including development-only scripts that you don't want to include in your production build, e.g. Angular's mock http service and a script that you use to configure it.